+++ penman/layout.py
@@ -516,23 +516,6 @@ def _rearrange(node: Node, key: Callable[[Branch], Any]) -> None:
     branches[:] = first + sorted(rest, key=key)
 
 
-def has_valid_layout(g: Graph,
-                     top: Variable = None,
-                     model: Model = None,
-                     strict: bool = False) -> bool:
-    """
-    Return ``True`` if *g* contains the information for a valid layout.
-
-    Having a valid layout means that the graph data allows a
-    depth-first traversal that reconstructs a spanning tree used for
-    serialization.
-    """
-    if model is None:
-        model = _default_model
-    tree, data, nodemap, variables = _configure(g, top, model, strict)
-    return len(data) == 0
-
-
 def get_pushed_variable(g: Graph,
                         triple: BasicTriple) -> Union[Variable, None]:
     """